requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1

fastapi_sso
argon2-cffi
python-jose[cryptography]
redis